Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,6 +3,6 @@
 int main () {
     int n ;
     scanf("%d",&n);
-    int area = n * n * 3.14 ;
-    printf("Area: %d",area);
+    float area = n * n * 3.14 ;
+    printf("Area: %f",area);
 }
